requirements.txt
amqp==5.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-types==0.7.0
anyio==4.8.0
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.8.2
boltons @ file:///croot/boltons_1677628692245/work
Brotli @ file:///croot/brotli-split_1714483155106/work
celery==5.3.4
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
conda @ file:///croot/conda_1714403036266/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1714483155348/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
cryptography==41.0.4
distro @ file:///croot/distro_1714488253808/work
exceptiongroup==1.2.2
Flask==2.3.3
Flask-Cors==4.0.1
groq==0.18.0
gunicorn==21.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
kombu==5.4.2
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
loguru==0.7.2
MarkupSafe==2.1.5
menuinst @ file:///croot/menuinst_1714510563922/work
openai==1.62.0
optionaldict==0.1.2
packaging @ file:///croot/packaging_1710807400464/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024445381/work
prompt_toolkit==3.0.50
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.0
requests @ file:///croot/requests_1707355572290/work
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
tenacity==9.0.0
tld==0.13
tqdm @ file:///croot/tqdm_1714567712644/work
typing_extensions==4.12.2
tzdata==2025.1
urllib3 @ file:///croot/urllib3_1707770551213/work
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
wechatpy==1.8.18
Werkzeug==3.0.3
xmltodict==0.14.2
zipp==3.19.1
zstandard @ file:///croot/zstandard_1714677652653/work
